gpuav: Rename interface to GpuShaderInstrumentor
