+++ layers/gpu_validation/gpu_state_tracker.h
@@ -15,19 +15,16 @@
  * limitations under the License.
  */
 #pragma once
-#include "generated/chassis.h"
-#include "gpu_validation/gpu_resources.h"
 #include "state_tracker/cmd_buffer_state.h"
 #include "state_tracker/queue_state.h"
-#include "vma/vma.h"
+
+class GpuShaderInstrumentor;
 
 namespace gpu_tracker {
 
-class Validator;
-
 class Queue : public vvl::Queue {
   public:
-    Queue(Validator &state, VkQueue q, uint32_t index, VkDeviceQueueCreateFlags flags,
+    Queue(GpuShaderInstrumentor &shader_instrumentor_, VkQueue q, uint32_t index, VkDeviceQueueCreateFlags flags,
           const VkQueueFamilyProperties &queueFamilyProperties);
     virtual ~Queue();
 
@@ -37,7 +34,7 @@ class Queue : public vvl::Queue {
     void SubmitBarrier(const Location &loc, uint64_t seq);
     void Retire(vvl::QueueSubmission &) override;
 
-    Validator &state_;
+    GpuShaderInstrumentor &shader_instrumentor_;
     VkCommandPool barrier_command_pool_{VK_NULL_HANDLE};
     VkCommandBuffer barrier_command_buffer_{VK_NULL_HANDLE};
     VkSemaphore barrier_sem_{VK_NULL_HANDLE};
@@ -46,8 +43,8 @@ class Queue : public vvl::Queue {
 
 class CommandBuffer : public vvl::CommandBuffer {
   public:
-    CommandBuffer(Validator &gpuav, VkCommandBuffer handle, const VkCommandBufferAllocateInfo *pCreateInfo,
-                  const vvl::CommandPool *pool);
+    CommandBuffer(GpuShaderInstrumentor &shader_instrumentor_, VkCommandBuffer handle,
+                  const VkCommandBufferAllocateInfo *pCreateInfo, const vvl::CommandPool *pool);
 
     virtual bool PreProcess() = 0;
     virtual void PostProcess(VkQueue queue, const Location &loc) = 0;
@@ -56,168 +53,3 @@ class CommandBuffer : public vvl::CommandBuffer {
 
 VALSTATETRACK_DERIVED_STATE_OBJECT(VkQueue, gpu_tracker::Queue, vvl::Queue)
 VALSTATETRACK_DERIVED_STATE_OBJECT(VkCommandBuffer, gpu_tracker::CommandBuffer, vvl::CommandBuffer)
-
-namespace gpu_tracker {
-
-class DescriptorSetManager {
-  public:
-    DescriptorSetManager(VkDevice device, uint32_t num_bindings_in_set);
-    ~DescriptorSetManager();
-
-    VkResult GetDescriptorSet(VkDescriptorPool *out_desc_pool, VkDescriptorSetLayout ds_layout, VkDescriptorSet *out_desc_sets);
-    VkResult GetDescriptorSets(uint32_t count, VkDescriptorPool *out_pool, VkDescriptorSetLayout ds_layout,
-                               std::vector<VkDescriptorSet> *out_desc_sets);
-    void PutBackDescriptorSet(VkDescriptorPool desc_pool, VkDescriptorSet desc_set);
-
-  private:
-    std::unique_lock<std::mutex> Lock() const { return std::unique_lock<std::mutex>(lock_); }
-
-    static const uint32_t kItemsPerChunk = 512;
-    struct PoolTracker {
-        uint32_t size;
-        uint32_t used;
-    };
-    VkDevice device;
-    uint32_t num_bindings_in_set;
-    vvl::unordered_map<VkDescriptorPool, struct PoolTracker> desc_pool_map_;
-    mutable std::mutex lock_;
-};
-
-struct GpuAssistedShaderTracker {
-    VkPipeline pipeline;
-    VkShaderModule shader_module;
-    VkShaderEXT shader_object;
-    std::vector<uint32_t> instrumented_spirv;
-};
-
-class Validator : public ValidationStateTracker {
-  public:
-    using BaseClass = ValidationStateTracker;
-
-    ReadLockGuard ReadLock() const override;
-    WriteLockGuard WriteLock() override;
-    void PreCallRecordCreateDevice(VkPhysicalDevice physicalDevice, const VkDeviceCreateInfo *pCreateInfo,
-                                   const VkAllocationCallbacks *pAllocator, VkDevice *pDevice, const RecordObject &record_obj,
-                                   vku::safe_VkDeviceCreateInfo *modified_create_info) override;
-    void CreateDevice(const VkDeviceCreateInfo *pCreateInfo, const Location &loc) override;
-    void PreCallRecordDestroyDevice(VkDevice device, const VkAllocationCallbacks *pAllocator,
-                                    const RecordObject &record_obj) override;
-
-    bool ValidateCmdWaitEvents(VkCommandBuffer command_buffer, VkPipelineStageFlags2 src_stage_mask, const Location &loc) const;
-    bool PreCallValidateCmdWaitEvents(VkCommandBuffer commandBuffer, uint32_t eventCount, const VkEvent *pEvents,
-                                      VkPipelineStageFlags srcStageMask, VkPipelineStageFlags dstStageMask,
-                                      uint32_t memoryBarrierCount, const VkMemoryBarrier *pMemoryBarriers,
-                                      uint32_t bufferMemoryBarrierCount, const VkBufferMemoryBarrier *pBufferMemoryBarriers,
-                                      uint32_t imageMemoryBarrierCount, const VkImageMemoryBarrier *pImageMemoryBarriers,
-                                      const ErrorObject &error_obj) const override;
-    bool PreCallValidateCmdWaitEvents2KHR(VkCommandBuffer commandBuffer, uint32_t eventCount, const VkEvent *pEvents,
-                                          const VkDependencyInfoKHR *pDependencyInfos, const ErrorObject &error_obj) const override;
-    bool PreCallValidateCmdWaitEvents2(VkCommandBuffer commandBuffer, uint32_t eventCount, const VkEvent *pEvents,
-                                       const VkDependencyInfo *pDependencyInfos, const ErrorObject &error_obj) const override;
-    void PreCallRecordCreatePipelineLayout(VkDevice device, const VkPipelineLayoutCreateInfo *pCreateInfo,
-                                           const VkAllocationCallbacks *pAllocator, VkPipelineLayout *pPipelineLayout,
-                                           const RecordObject &record_obj, chassis::CreatePipelineLayout &chassis_state) override;
-    void PostCallRecordCreatePipelineLayout(VkDevice device, const VkPipelineLayoutCreateInfo *pCreateInfo,
-                                            const VkAllocationCallbacks *pAllocator, VkPipelineLayout *pPipelineLayout,
-                                            const RecordObject &record_obj) override;
-    void PreCallRecordCreateShadersEXT(VkDevice device, uint32_t createInfoCount, const VkShaderCreateInfoEXT *pCreateInfos,
-                                       const VkAllocationCallbacks *pAllocator, VkShaderEXT *pShaders,
-                                       const RecordObject &record_obj, chassis::ShaderObject &chassis_state) override;
-    void PostCallRecordCreateShadersEXT(VkDevice device, uint32_t createInfoCount, const VkShaderCreateInfoEXT *pCreateInfos,
-                                        const VkAllocationCallbacks *pAllocator, VkShaderEXT *pShaders,
-                                        const RecordObject &record_obj, chassis::ShaderObject &chassis_state) override;
-    void PreCallRecordDestroyShaderEXT(VkDevice device, VkShaderEXT shader, const VkAllocationCallbacks *pAllocator,
-                                       const RecordObject &record_obj) override;
-
-    void PreCallRecordCreateGraphicsPipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                              const VkGraphicsPipelineCreateInfo *pCreateInfos,
-                                              const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                              const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                              chassis::CreateGraphicsPipelines &chassis_state) override;
-    void PreCallRecordCreateComputePipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                             const VkComputePipelineCreateInfo *pCreateInfos,
-                                             const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                             const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                             chassis::CreateComputePipelines &chassis_state) override;
-    void PreCallRecordCreateRayTracingPipelinesNV(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                                  const VkRayTracingPipelineCreateInfoNV *pCreateInfos,
-                                                  const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                  const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                                  chassis::CreateRayTracingPipelinesNV &chassis_state) override;
-    void PreCallRecordCreateRayTracingPipelinesKHR(VkDevice device, VkDeferredOperationKHR deferredOperation,
-                                                   VkPipelineCache pipelineCache, uint32_t count,
-                                                   const VkRayTracingPipelineCreateInfoKHR *pCreateInfos,
-                                                   const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                   const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                                   chassis::CreateRayTracingPipelinesKHR &chassis_state) override;
-    void PostCallRecordCreateGraphicsPipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                               const VkGraphicsPipelineCreateInfo *pCreateInfos,
-                                               const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                               const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                               chassis::CreateGraphicsPipelines &chassis_state) override;
-    void PostCallRecordCreateComputePipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                              const VkComputePipelineCreateInfo *pCreateInfos,
-                                              const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                              const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                              chassis::CreateComputePipelines &chassis_state) override;
-    void PostCallRecordCreateRayTracingPipelinesNV(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
-                                                   const VkRayTracingPipelineCreateInfoNV *pCreateInfos,
-                                                   const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                   const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                                   chassis::CreateRayTracingPipelinesNV &chassis_state) override;
-    void PostCallRecordCreateRayTracingPipelinesKHR(VkDevice device, VkDeferredOperationKHR deferredOperation,
-                                                    VkPipelineCache pipelineCache, uint32_t count,
-                                                    const VkRayTracingPipelineCreateInfoKHR *pCreateInfos,
-                                                    const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                    const RecordObject &record_obj, PipelineStates &pipeline_states,
-                                                    chassis::CreateRayTracingPipelinesKHR &chassis_state) override;
-    void PreCallRecordDestroyPipeline(VkDevice device, VkPipeline pipeline, const VkAllocationCallbacks *pAllocator,
-                                      const RecordObject &record_obj) override;
-
-    void ReportSetupProblem(LogObjectList objlist, const Location &loc, const char *const specific_message,
-                            bool vma_fail = false) const;
-    bool CheckForGpuAvEnabled(const void *pNext);
-
-  protected:
-    std::shared_ptr<vvl::Queue> CreateQueue(VkQueue q, uint32_t index, VkDeviceQueueCreateFlags flags,
-                                            const VkQueueFamilyProperties &queueFamilyProperties) override;
-
-    template <typename CreateInfo, typename SafeCreateInfo, typename ChassisState>
-    void PreCallRecordPipelineCreations(uint32_t count, const CreateInfo *pCreateInfos, const VkAllocationCallbacks *pAllocator,
-                                        VkPipeline *pPipelines, PipelineStates &pipeline_states,
-                                        std::vector<SafeCreateInfo> *new_pipeline_create_infos, const RecordObject &record_obj,
-                                        ChassisState &chassis_state);
-    template <typename CreateInfo, typename SafeCreateInfo>
-    void PostCallRecordPipelineCreations(const uint32_t count, const CreateInfo *pCreateInfos,
-                                         const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                         const SafeCreateInfo &modified_create_infos);
-
-    virtual bool InstrumentShader(const vvl::span<const uint32_t> &input, std::vector<uint32_t> &instrumented_spirv,
-                                  uint32_t unique_shader_id, const Location &loc) = 0;
-
-  public:
-    mutable bool aborted = false;
-    bool force_buffer_device_address;
-    vvl::unordered_map<uint32_t, std::pair<size_t, std::vector<uint32_t>>> instrumented_shaders;
-    PFN_vkSetDeviceLoaderData vkSetDeviceLoaderData;
-    const char *setup_vuid;
-    VkPhysicalDeviceFeatures supported_features{};
-    VkPhysicalDeviceFeatures desired_features{};
-    uint32_t adjusted_max_desc_sets = 0;
-    std::atomic<uint32_t> unique_shader_module_id = 1;  // zero represents no shader module found
-    uint32_t output_buffer_byte_size = 0;
-    VkDescriptorSetLayout debug_desc_layout_ = VK_NULL_HANDLE;
-    // This is a layout used to "pad" a pipeline layout to fill in any gaps to the selected bind index
-    VkDescriptorSetLayout dummy_desc_layout_ = VK_NULL_HANDLE;
-    VkPipelineLayout debug_pipeline_layout = VK_NULL_HANDLE;
-    uint32_t desc_set_bind_index = 0;
-    VmaAllocator vmaAllocator = {};
-    VmaPool output_buffer_pool = VK_NULL_HANDLE;
-    std::unique_ptr<DescriptorSetManager> desc_set_manager;
-    vvl::concurrent_unordered_map<uint32_t, GpuAssistedShaderTracker> shader_map;
-    std::vector<VkDescriptorSetLayoutBinding> validation_bindings_;
-
-    gpuav::DeviceMemoryBlock indices_buffer{};
-};
-
-}  // namespace gpu_tracker
